Update apA.md
I couldn't find toMethod() in any documentation. According to the stackoverflow-user Bergi "the toMethod method was experimental only and did not make it into the standard. The home object is essentially static now, the only way to manipulate super is to change the [[prototype]]" (https://stackoverflow.com/questions/27511704/what-does-function-prototype-tomethod-do). This could explain the fact that in the original code the call of the toMethod() broke the code.